- Se implementa la lista paginada y filtrada de los usuarios registrados con solicitudes de prestamo.

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/solicitudescrediya/r2dbc/typeReactiveRepository/TypeReactiveRepository.java
@@ -14,7 +14,7 @@ public interface TypeReactiveRepository extends ReactiveCrudRepository<LoanTypeE
         ReactiveQueryByExampleExecutor<LoanTypeEntity> {
 
     @Query("SELECT * FROM tipo_prestamo WHERE id_tipo_prestamo = :typeLoanId")
-    Mono<LoanTypeEntity> existLoanTypeById(int typeLoanId);
+    Mono<LoanTypeEntity> existLoanTypeById(long typeLoanId);
 
     @Query("SELECT * FROM tipo_prestamo " +
             "WHERE id_tipo_prestamo = :idType " +
